rename radial classes and packages
