+++ src/main/java/de/teamlapen/vampirism/client/gui/screens/EditSelectMinionTaskScreen.java
@@ -7,7 +7,7 @@ import de.teamlapen.vampirism.api.entity.minion.IFactionMinionTask;
 import de.teamlapen.vampirism.api.entity.minion.INoGlobalCommandTask;
 import de.teamlapen.vampirism.api.util.ItemOrdering;
 import de.teamlapen.vampirism.client.ClientConfigHelper;
-import de.teamlapen.vampirism.client.gui.screens.radial.ReorderingGuiRadialMenu;
+import de.teamlapen.vampirism.client.gui.screens.radial.edit.ReorderingGuiRadialMenu;
 import de.teamlapen.vampirism.entity.factions.FactionPlayerHandler;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
